feat: added hotkeys
I added the hotkey events for the user dropdown menu.
