+++ apps/nextjs/src/app/dashboard/layout.tsx
@@ -1,42 +1,15 @@
-"use client";
-
-import { useState } from "react";
 import Image from "next/image";
 import Link from "next/link";
 
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from "@acme/ui/dropdown-menu";
+import HotKeyEventListeners from "../_components/navbar/HotKeyEventListeners";
+import NavProjectDropdown from "../_components/navbar/NavProjectDropdown";
+import NavUserDropdown from "../_components/navbar/NavUserDropdown";
 
 export default function GlobalNavbarLayout(props: {
   children: React.ReactNode;
 }) {
-  /**
-   * This is a mockup of the data that would be passed to the Navbar component just so that we can see how it would look.
-   */
-  const username = "Ben Davies";
-
-  interface Project {
-    projectName: string;
-    projectIcon: string;
-  }
-
-  const projects: Project[] = [
-    { projectName: "Project 1", projectIcon: "projectImage1" },
-    { projectName: "Project 2", projectIcon: "projectImage2" },
-    { projectName: "Project 3", projectIcon: "projectImage3" },
-  ];
-  // This is the current project state that is being displayed in the project dropdown
-  const [currentProjectState, setCurrentProjectState] = useState(
-    projects[0]?.projectName ?? "",
-  );
-
   return (
-    <div>
+    <>
       <div className="flex flex-row justify-between gap-4 px-12 py-8">
         {/* This is the Pump Logo */}
         <div className="flex flex-row items-center justify-center gap-8">
@@ -46,6 +19,7 @@ export default function GlobalNavbarLayout(props: {
               alt="Chevron Down"
               width={26}
               height={18}
+              className="h-auto w-auto"
             />
             <h1 className="text-2xl font-bold">pump.task</h1>
           </div>
@@ -54,129 +28,11 @@ export default function GlobalNavbarLayout(props: {
           </h5>
         </div>
 
-        {/* This is the Project dropdown Menu. Quick note will probably make this conditional rendered in the future so if users don't have any projects it just renders a button to create a new project. */}
         <div className="flex gap-10 hover:cursor-pointer">
-          <DropdownMenu>
-            <DropdownMenuTrigger className="relative flex cursor-default select-none items-center rounded-md border px-4 py-2 text-sm outline-none transition-colors hover:cursor-pointer hover:bg-accent hover:text-accent-foreground focus:bg-accent focus:text-accent-foreground">
-              <div className="flex flex-row items-center gap-4">
-                <Image
-                  className="inline-block h-4 w-4 rounded-full"
-                  src="/badge.png"
-                  alt="badge"
-                  width={12}
-                  height={12}
-                />
-                <h6 className="text-sm">{currentProjectState}</h6>
-                <Image
-                  src="/chevron-down.svg"
-                  alt="Chevron Down"
-                  width={16}
-                  height={16}
-                />
-              </div>
-            </DropdownMenuTrigger>
-
-            {/* This is the map function to display the users projects inside the projects drop down menu*/}
-            <DropdownMenuContent>
-              {projects.map((project, index) => (
-                <DropdownMenuItem
-                  className="flex flex-row items-center gap-4 hover:cursor-pointer"
-                  key={index}
-                  onClick={() => setCurrentProjectState(project.projectName)}
-                >
-                  <div className="flex flex-row items-center gap-4 hover:cursor-pointer">
-                    <Image
-                      className="inline-block h-5 w-5 rounded-full"
-                      src="/badge.png"
-                      alt="badge"
-                      width={20}
-                      height={20}
-                    />
-                    <h1 className="text-sm">{project.projectName}</h1>
-                  </div>
-                </DropdownMenuItem>
-              ))}
-              <DropdownMenuSeparator />
-              <DropdownMenuItem className="hover:cursor-pointer">
-                Create New Project
-              </DropdownMenuItem>
-            </DropdownMenuContent>
-          </DropdownMenu>
-
-          {/* This is the users Drop down Menu */}
-          <DropdownMenu>
-            <DropdownMenuTrigger className="relative flex cursor-default select-none items-center rounded-md border px-4 py-2 text-sm outline-none transition-colors hover:cursor-pointer hover:bg-accent hover:text-accent-foreground focus:bg-accent focus:text-accent-foreground">
-              <div className="flex flex-row items-center gap-4">
-                <Image
-                  className="inline-block h-4 w-4 rounded-full"
-                  src="/userProfileIcon.png"
-                  alt="badge"
-                  width={12}
-                  height={12}
-                />
-                <h6 className="text-sm">{username}</h6>
-                <Image
-                  src="/chevron-down.svg"
-                  alt="Chevron Down"
-                  width={16}
-                  height={16}
-                />
-              </div>
-            </DropdownMenuTrigger>
-            <DropdownMenuContent>
-              <DropdownMenuItem className="flex flex-row justify-between hover:cursor-pointer">
-                <div className="flex flex-row gap-4">
-                  <Image
-                    src="/userIcon.png"
-                    alt="userIcon"
-                    width={20}
-                    height={20}
-                  />
-                  <h5>Profile</h5>
-                </div>
-                <Image
-                  src="/ProfileShortcut.png"
-                  alt="Profile"
-                  width={32}
-                  height={32}
-                />
-              </DropdownMenuItem>
-              <DropdownMenuItem className="flex flex-row justify-between gap-16 hover:cursor-pointer">
-                <div className="flex flex-row gap-4">
-                  <Image
-                    src="/SettingsIcon.png"
-                    alt="Settings"
-                    width={20}
-                    height={20}
-                  />
-                  <h5>Settings</h5>
-                </div>
-                <Image
-                  src="/SettingShortcut.png"
-                  alt="Settings"
-                  width={20}
-                  height={20}
-                />
-              </DropdownMenuItem>
-              <DropdownMenuItem className="flex flex-row justify-between gap-16 hover:cursor-pointer">
-                <div className="flex flex-row gap-4">
-                  <Image
-                    src="/LogoutIcon.png"
-                    alt="Chevron Down"
-                    width={20}
-                    height={20}
-                  />
-                  <h5>Log out</h5>
-                </div>
-                <Image
-                  src="/LogoutShortcut.png"
-                  alt="Chevron Down"
-                  width={32}
-                  height={32}
-                />
-              </DropdownMenuItem>
-            </DropdownMenuContent>
-          </DropdownMenu>
+          {/* This is the Project dropdown Menu. Quick note will probably make this conditional rendered in the future so if users don't have any projects it just renders a button to create a new project. */}
+          <NavProjectDropdown />
+          {/* This is the users Drop down Menu with profile, setting and logout buttons*/}
+          <NavUserDropdown />
         </div>
       </div>
 
@@ -214,8 +70,11 @@ export default function GlobalNavbarLayout(props: {
         </Link>
       </div>
 
+      {/* This is the event listener for hotkeys profile, setting, logout */}
+      <HotKeyEventListeners />
+
       {/* This just specifies where the children components will be rendered in application */}
       {props.children}
-    </div>
+    </>
   );
 }
