+++ apps/nextjs/src/app/dashboard/page.tsx
@@ -1,7 +1,7 @@
 export default function DashboardPage() {
   return (
     <div>
-      <h1>Dashboard</h1>
+      <h1>Dashboard / Home Page</h1>
     </div>
   );
 }
